Add custom categorical_to_tf_graph
Fix categorical to tf graph for classification problem.

For a binary class label, there will be two categorical leaf nodes in the SPN.
One which one-hot encodes the first class as [0, 1] and one that encodes the
second clas as [1, 0].

Since the tf optimizes the log likelihood, these one-hot represented
probabilities will be projected into logspace which results in
log([1,0])=[0, -inf] and therefore NaNs in further computations.

Therefore, this custom method adds a small epsilon, such that the zero
probability value in the one-hot vector will not degrade to negative infinity.