🏗️ __version__ information
Add `__version__ = "0.0.40"` variable in `src/spn/_meta.py`
Start pulling version info in `setup.py` from `_meta`

The current documentation and `setup.py` had to be updated
in two places, which could lead to inconsistencies. Introducing
the variable here allows it to be updated once.

There is some complexity in how `setup.py` accesses the
__version__ variable. This is intended to prevent side-effects
potentially caused by importing the library during setup,
and is addressed in an in-line comment.


Former-commit-id: 6bc848b7f97066a5c5e02fb857aafcb2ef0a2ac1