⬆️ Adding `docs/requirements.txt` for documentation
Add `docs/requirements.txt`

Documentation has a specific set of dependencies which are
independent of the rest of the library. This `requirements.txt`
can be used to install these separately (e.g. in an automatic
build system, or for a user who does not need to install
the documentation libraries to use the code).