📝 Reworking API documentation
Move `api.rst` -> `source/documentation/api.rst`
Add `developing.rst` discussing notes on testing, docs
Start enumearting functions/classes as needed

This changes the API documentation quite a bit. The previous
version would use `sphinx-apidoc` to generate pages for
*everything* in the `src/` directory, which made it a little
difficult to navigate and know what to look for. **Instead**,
this lists the modules mentioned in the `examples/` and sorts
them into whether they deal with "Structure," "Learning,"
"Inference," "Utilities," or "Datasets."

Many of the modules are currently lacking docstrings, so
listing modules in this manner could also help focus on
which modules to document in the near future.