layerwise: Do marg. before leaf forward pass
This fixes an issue where, when in a gradient tracking context, NaN
inputs which will be marginalized anyway will lead to NaN loss since the
marginalization happened after the distribution forward pass and thus
does not break the backward chain between the NaN values and the rest.

The wrong ordering was likely a mistake by accident without further
notice until now.


Former-commit-id: c8ea5dc1c64c6728659c8e20530f4f09571e71f8