Fix #89
Fix keyword argument `clip_on` that behaved differently for 2.4/2.5

In `networkx==2.4`, `clip_on` could be passed as a keyword argument,
but it was unused. In `networkx==2.5` arbitrary keyword arguments
appear to no longer be passed. In the prerelease for
`networkx==2.6`, `clip_on` appears to be added back in as an explicit
parameter.

This drops the keyword argument to avoid these problems.


Former-commit-id: 5f7bc756d137b0719025e5e1c419c428611d8d57