Add standardization, alpha and number of clusters.
Add options for the clustering (standardization before k-means and the number of clusters per iteration) and for the histograms (whether to use alpha smoothing/Laplace correction). The number of clusters are created so that a two-layered MSPN can be created at once, with as many clusters as desired (instead of recursive clustering in two).

Former-commit-id: 34486ccc661843a8daa982aed8c2933aa6b40be7