Begin of major restructuring
This restructuring gets rid of the backend separation in leaf
implementations as well as the tensorly dependency. With a common leaf
implementation for all required backends we can streamline leaf node
implementations, reducing code size and test size massively. During this
restructuring we found that most of our required tensor operations were
not supported by tensorly or behaved weirdly across different backends.
Therefore, this change also drop the tensorly dependency and introduces
the `spflow.tensor` module in which we collect all operations
`spflow.tensor.ops` that are necessary for the computations any module,
as well as the backend handling `spflow.tensor.backend`. The
idea is heavily inspired by tensorly, but really cut down to only the
things we need in spflow. With this we have full control over the tensor API.

Resolves: #151,#130,#105